Moves DPC struct into separate file
